% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.R
\docType{package}
\name{cate-package}
\alias{cate-package}
\title{High dimensional factor analysis and confounder adjusted testing and estimation (CATE)}
\description{
Provides several methods for factor analysis in high dimension (both n,p >> 1) and methods to adjust for possible confounders in multiple hypothesis testing.
}
\seealso{
\code{\link{factor.analysis}}, \code{\link{cate}}
}
