% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.R
\name{check.rank}
\alias{check.rank}
\title{Check linear dependence}
\usage{
check.rank(X.primary, X.nuis)
}
\description{
Checks if \code{X.primary} has full rank and is linearly independent of \code{x.nuis}.
}
\details{
X.nuis is allowed to not have full rank.
}
\keyword{internal}
