% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_functions.R
\name{EigenDiff}
\alias{EigenDiff}
\title{the ed method to estimate the number of factors proposed by Onatski(2010)}
\usage{
EigenDiff(Y, rmax = 20, niter = 10)
}
\arguments{
\item{Y}{outcome, n*p matrix}

\item{rmax}{the maximum number of factors to consider. If the estimated number of factors is rmax,
then users are encouraged to increase rmax and run again. Default is 20.}
}
\description{
the ed method to estimate the number of factors proposed by Onatski(2010)
}
\keyword{internal}
