% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_functions.R
\name{fa.em}
\alias{fa.em}
\title{Factor analysis via EM algorithm to maximize likelihood}
\usage{
fa.em(Y, r, tol = 1e-06, maxiter = 1000)
}
\arguments{
\item{Y}{data matrix, a n*p matrix}

\item{r}{number of factors}

\item{tol}{a tolerance scale of change of log-likelihood for convergence in the EM iterations}

\item{maxiter}{maximum iterations}
}
\description{
Factor analysis via EM algorithm to maximize likelihood
}
\references{
{
Bai, J. and Li, K. (2012). Statistical analysis of factor models of high dimension. \emph{The Annals of Statistics 40}, 436-465.
}
}
\seealso{
\code{\link{factor.analysis}} for the main function.
}
