% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_functions.R
\name{fa.pc}
\alias{fa.pc}
\title{Factor analysis via principal components}
\usage{
fa.pc(Y, r)
}
\arguments{
\item{Y}{data matrix, a n*p matrix}

\item{r}{number of factors}
}
\description{
Factor analysis via principal components
}
\seealso{
\code{\link{factor.analysis}} for the main function.
}
