% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate.R
\name{parse.cate.formula}
\alias{parse.cate.formula}
\title{Parse the formula in \code{cate}, and return a matrix of primary variables and a matrix of nuisance variables.}
\usage{
parse.cate.formula(formula, X.data = NULL)
}
\description{
Parse the formula in \code{cate}, and return a matrix of primary variables and a matrix of nuisance variables.
}
\keyword{internal}
